Fix compiler warnings in DirectX11 sources
Clears every warning MSVC reports for our own sources (DirectXTK and
SDK headers excluded) so new ones stand out:

- DEFINE_OPERATOR bodies return int/double from a float evaluate();
  the conversion is the operator contract, made explicit with one
  cast in the macro (C4244 x12).
- %d used with size_t in SetDataCacheRegion logs (C4477).
- GetIniInt called with -1.0f defaults for int settings.
- Explicit casts at the remaining narrowing/sign-mismatch sites
  (size_t/UINT64/uint64 -> int/UINT, std::distance -> int, sentinel
  -1 comparisons) and TargetInfo::max_slot_count made unsigned.
- Three unreferenced locals removed, _wfopen -> _wfopen_s.
- GetElementIndex: default: assert(false) fell off the end in release
  builds; return SIZE_MAX (C4715).
- KeyOverride gets the same C4250 suppress KeyOverrideCycle has.

+++ DirectX11/HackerDXGI.cpp
@@ -558,11 +558,11 @@ STDMETHODIMP HackerSwapChain::Present(THIS_
 			if (G->overlay_buffer_hash_lifetime >= 0)
 				PurgeStaleVisitedBufferHashes(mHackerDevice);
 			if (G->mSelectedIndexBufferPos == INT_MAX) {
-				G->mSelectedIndexBufferPos = G->mVisitedIndexBuffers.size() - 1;
+				G->mSelectedIndexBufferPos = (int)G->mVisitedIndexBuffers.size() - 1;
 				G->mSelectedIndexBuffer = *std::prev(G->mVisitedIndexBuffers.end());
 			}
 			if (G->mSelectedVertexBufferPos == INT_MAX) {
-				G->mSelectedVertexBufferPos = G->mVisitedVertexBuffers.size() - 1;
+				G->mSelectedVertexBufferPos = (int)G->mVisitedVertexBuffers.size() - 1;
 				G->mSelectedVertexBuffer = *std::prev(G->mVisitedVertexBuffers.end());
 			}
 			if (G->gResetSelectedVertexBufferSlotId) {
